+++ code/crates/core-consensus/src/handle/sync.rs
@@ -1,4 +1,4 @@
-use crate::handle::decide::decide;
+use crate::handle::decide::{decide, decide_current_no_timeout};
 use crate::handle::driver::apply_driver_input;
 use crate::handle::signature::verify_certificate;
 use crate::handle::validator_set::get_validator_set;
@@ -43,19 +43,7 @@ where
     .await?;
 
     if state.driver.step_is_commit() {
-        perform!(
-            co,
-            Effect::CancelTimeout(Timeout::commit(state.driver.round()), Default::default())
-        );
-
-        let height = state.driver.height();
-        let round = state.driver.round();
-        let proposal = state
-            .decision
-            .remove(&(height, round))
-            .ok_or_else(|| Error::DecidedValueNotFound(height, round))?;
-
-        decide(co, state, metrics, round, proposal).await?;
+        decide_current_no_timeout(co, state, metrics).await?;
     }
 
     Ok(())
